docs: add a bunch of comments
