+++ resources/js/dynamic/reorderSiteAwards.ts
@@ -1,2 +1,3 @@
 export * from './reorderSiteAwards/index';
 export * from './reorderSiteAwards/handleHideSavedHiddenRowsChange';
+export * from './reorderSiteAwards/handleRowDragOver';
